for #601, refactor the single base integrate test cases
