+++ sharding-jdbc/src/test/java/io/shardingsphere/dbtest/engine/SingleIntegrateTest.java
@@ -60,17 +60,7 @@ public abstract class SingleIntegrateTest extends BaseIntegrateTest {
         sql = getSQL(sqlCaseId);
         expectedDataFile = getExpectedDataFile(path, shardingRuleType, databaseTypeEnvironment.getDatabaseType(), assertion.getExpectedDataFile());
     }
-    
-    @BeforeClass
-    public static void initDatabasesAndTables(){
-        createDatabasesAndTables();
-    }
-    
-    @AfterClass
-    public static void destroyDatabasesAndTables(){
-        dropDatabases();
-    }
-    
+
     private String getSQL(final String sqlCaseId) throws ParseException {
         List<String> parameters = new LinkedList<>();
         for (SQLValue each : assertion.getSQLValues()) {
