+++ sharding-jdbc/src/test/java/io/shardingsphere/dbtest/engine/dql/BaseDQLIntegrateTest.java
@@ -63,9 +63,10 @@ public abstract class BaseDQLIntegrateTest extends SingleIntegrateTest {
                                 final DatabaseTypeEnvironment databaseTypeEnvironment, final SQLCaseType caseType) throws IOException, JAXBException, SQLException, ParseException {
         super(sqlCaseId, path, assertion, shardingRuleType, databaseTypeEnvironment, caseType);
     }
-    
+
     @BeforeClass
     public static void insertData() throws IOException, JAXBException, SQLException, ParseException {
+        createDatabasesAndTables();
         for (DatabaseType each : DatabaseType.values()) {
             if (IntegrateTestEnvironment.getInstance().getDatabaseTypes().contains(each)) {
                 insertData(each);
@@ -86,6 +87,7 @@ public abstract class BaseDQLIntegrateTest extends SingleIntegrateTest {
                 clearData(each);
             }
         }
+        dropDatabases();
     }
     
     private static void clearData(final DatabaseType databaseType) throws SQLException, IOException, JAXBException {
